fix: address issue with top-level promise
app.start() is an async call and therefore returns promise. According to
the configured eslint strict rules it needs to be awaited. At the same
time top level await requires switching to ES6 modules in ts output. As
I'd like to keep it commonjs at this stage I've decided to address the
problem in a more straightforward manner: simply by awaiting the promise
in the top level server module using the classic Promise/then/catch
idiom.