feat: add typing to the route handlers
Reverted to the inline definition of the handler function: realized that
with the current level of ts knowledge I'm unable to write the correct
generic defineHandler() call (commented the lines causing tsc error in
it).

This commit basically completes the typing of the code: no errors
reported by tsc in strict mode and the tests are passed as well.
There're still several cycle dependencies reported by eslint though
which need to be addressed.