feat: move Id generation out of database
I've decided that it'd be preferable to generate Ids of the objects
stored in the database externally and store the Id along in the object
itself along the other field.

This also allows to treat the Db as a generic key-value storage with the
type of the key passed by user. It also decouples the dependencies
between the modules therefore removing cycle dependencies.
It's true that there's now some duplication however: the Id is stored
both as the key of the Map and the value (the in the property of the
respective object). But the key of the Map can be conceptually treated
as the "index" so from this perspective the "duplication" is natural.