fix: rollback to ES2021 version of JS code
The point is that the remote eslint config used in github workflow has
parser option "ecmaVersion": "2021". It means that the latest JS
language features like class static properties are not recognized
therefore resulting in linter errors which I'd like to avoid. So I
rolled back the local eslint config to use the same ecmaVersion and also
changed the code a bit to accommodate to its limitations.