docs: configure typedoc
For some reason the providing just the entry point src/server.ts doesn't
make typedoc to generate any html docs for some reason (TODO: try to
figure out what's that reason of that). Changing the entryPointStrategy
to "expand" and passing the src/ directory as the entry point seems to
solve the issue and the docs are generated (TODO: however, the page for
the same src/server.ts is missing in this case for some reason).

https://typedoc.org/guides/options/#input-options

Also configured the npm script to re-generate the docs with nodemon.
Tip: it's quite handy to use it together with Live Server extension.

https://marketplace.visualstudio.com/items?itemName=ritwickdey.LiveServer