feat: pass local server bind address
We need to bind to local address 0.0.0.0 (not 127.0.0.1 which is the
default) inside docker container to be able to expose the container
server port to be available from outside. For that we need to pass the
additional argument to fastify.listen() call. The value of that argument
is taken from the environment variable ADDR which now lives alongside
PORT.