feat: establish database connection
Added the logic that connects to the database using the provide
connection options. However haven't yet switched the logic to store the
actual data in/retrieve it from the database. So at this point both
in-memory DB and postgres connection co-exist in the same server
instance (the tests do pass).

The error handling happens with the help of the new custom exception
classes.