feat: consider order on entity list output
When the list of the objects of certain type is requested they should
preferably be sorted on their 'order' field if it's present in the
schema (it is for Task and BoardColumn).
Also disabled typeorm logging.

TODO: seems that the default sort order defined for an entity doesn't
actually work as declared so board_column records are not sorted by
'order'
https://github.com/typeorm/typeorm/blob/master/sample/sample30-default-order-by/entity/Post.ts#L6