fix: sort board columns by 'order' on output
Otherwise one of the tests ('should get a board by id') sometimes
"flashes": it expects that the columns of the board are returned in
order. But the simple select from board table with the left join on
border_column doesn't guarantee that as it turns out. It wasn't possible
to address this either with the relation decorator options or Typeorm
repository find() method so I had to use Typeorm query builder to
generate the correct SQL with necessary ORDER BY clause.

https://typeorm.io/#select-query-builder/joining-relations