fix: hash password on update request (PUT)
To deal with this correctly I had to change the schema and make login
and password fields of both POST and PUT requests mandatory.

I wouldn't like to change the logic of handling updates globally: to
merge the fields of the existing and the newly arrived object so I'm
just making all the fields mandatory to guarantee that they are replaced
together.