feat: expose columnId on /columns route only
Using 'group' serializer option we can control when the certain fields
of the entity are returned. In this case we don't want to return the
column Id when the column is returned as a part of a board but do want
to return the Id if the column is requested via its own route.

https://gregornovak.eu/nestjs-response-groups-serialization