feat: add logger interceptor
It can potentially be used to log request body in Fastify mode (the
problem is that it's not yet available when the regular pino logger
serializer is called).

But I'm still considering the better ways to approach this. Ideally
there should be a single place responsible for logging request, reply
and response time.